Rename SPI spooling package to spool
