+++ core/trino-main/src/main/java/io/trino/server/protocol/spooling/CoordinatorSegmentResource.java
@@ -20,8 +20,8 @@ import io.trino.server.ExternalUriInfo;
 import io.trino.server.protocol.spooling.SpoolingConfig.SegmentRetrievalMode;
 import io.trino.server.security.ResourceSecurity;
 import io.trino.spi.HostAddress;
-import io.trino.spi.protocol.SpooledSegmentHandle;
-import io.trino.spi.protocol.SpoolingManager;
+import io.trino.spi.spool.SpooledSegmentHandle;
+import io.trino.spi.spool.SpoolingManager;
 import jakarta.ws.rs.GET;
 import jakarta.ws.rs.Path;
 import jakarta.ws.rs.PathParam;
