+++ core/trino-main/src/main/java/io/trino/server/protocol/spooling/SpoolingManagerBridge.java
@@ -17,10 +17,10 @@ import com.google.inject.Inject;
 import io.airlift.slice.Slice;
 import io.trino.server.protocol.spooling.SpoolingConfig.SegmentRetrievalMode;
 import io.trino.spi.TrinoException;
-import io.trino.spi.protocol.SpooledLocation;
-import io.trino.spi.protocol.SpooledSegmentHandle;
-import io.trino.spi.protocol.SpoolingContext;
-import io.trino.spi.protocol.SpoolingManager;
+import io.trino.spi.spool.SpooledLocation;
+import io.trino.spi.spool.SpooledSegmentHandle;
+import io.trino.spi.spool.SpoolingContext;
+import io.trino.spi.spool.SpoolingManager;
 import jakarta.ws.rs.ServiceUnavailableException;
 
 import javax.crypto.Cipher;
@@ -37,9 +37,9 @@ import java.util.Optional;
 import static io.airlift.slice.Slices.utf8Slice;
 import static io.airlift.slice.Slices.wrappedBuffer;
 import static io.trino.spi.StandardErrorCode.CONFIGURATION_INVALID;
-import static io.trino.spi.protocol.SpooledLocation.CoordinatorLocation;
-import static io.trino.spi.protocol.SpooledLocation.DirectLocation;
-import static io.trino.spi.protocol.SpooledLocation.coordinatorLocation;
+import static io.trino.spi.spool.SpooledLocation.CoordinatorLocation;
+import static io.trino.spi.spool.SpooledLocation.DirectLocation;
+import static io.trino.spi.spool.SpooledLocation.coordinatorLocation;
 import static java.util.Base64.getUrlDecoder;
 import static java.util.Base64.getUrlEncoder;
 import static java.util.Objects.requireNonNull;
