+++ core/trino-spi/pom.xml
@@ -217,6 +217,19 @@
                                     <old>method io.trino.spi.connector.RecordCursor io.trino.spi.connector.SystemTable::cursor(io.trino.spi.connector.ConnectorTransactionHandle, io.trino.spi.connector.ConnectorSession, io.trino.spi.predicate.TupleDomain&lt;java.lang.Integer&gt;, java.util.Set&lt;java.lang.Integer&gt;)</old>
                                     <new>method io.trino.spi.connector.RecordCursor io.trino.spi.connector.SystemTable::cursor(io.trino.spi.connector.ConnectorTransactionHandle, io.trino.spi.connector.ConnectorSession, io.trino.spi.predicate.TupleDomain&lt;java.lang.Integer&gt;, java.util.Set&lt;java.lang.Integer&gt;, io.trino.spi.connector.ConnectorSplit)</new>
                                 </item>
+                                <item>
+                                    <ignore>true</ignore>
+                                    <code>java.method.returnTypeTypeParametersChanged</code>
+                                    <old>method java.lang.Iterable&lt;io.trino.spi.protocol.SpoolingManagerFactory&gt; io.trino.spi.Plugin::getSpoolingManagerFactories()</old>
+                                    <new>method java.lang.Iterable&lt;io.trino.spi.spool.SpoolingManagerFactory&gt; io.trino.spi.Plugin::getSpoolingManagerFactories()</new>
+                                    <justification>Spooling package renamed</justification>
+                                </item>
+                                <item>
+                                    <ignore>true</ignore>
+                                    <code>java.class.removed</code>
+                                    <old>interface io.trino.spi.protocol.SpoolingManager</old>
+                                    <justification>Spooling package renamed</justification>
+                                </item>
                             </differences>
                         </revapi.differences>
                     </analysisConfiguration>
