+++ core/trino-spi/src/main/java/io/trino/spi/spool/SpooledLocation.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.spi.protocol;
+package io.trino.spi.spool;
 
 import io.airlift.slice.Slice;
 import io.trino.spi.Experimental;
