+++ core/trino-spi/src/main/java/io/trino/spi/spool/SpoolingContext.java
@@ -11,27 +11,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.spi.protocol;
+package io.trino.spi.spool;
 
-import io.opentelemetry.api.OpenTelemetry;
-import io.opentelemetry.api.trace.Tracer;
 import io.trino.spi.Experimental;
+import io.trino.spi.QueryId;
+
+import static java.util.Objects.requireNonNull;
 
 @Experimental(eta = "2025-05-31")
-public interface SpoolingManagerContext
+public record SpoolingContext(String encoding, QueryId queryId, long rows, long size)
 {
-    default OpenTelemetry getOpenTelemetry()
+    public SpoolingContext
     {
-        throw new UnsupportedOperationException();
-    }
-
-    default Tracer getTracer()
-    {
-        throw new UnsupportedOperationException();
-    }
-
-    default boolean isCoordinator()
-    {
-        throw new UnsupportedOperationException();
+        requireNonNull(queryId, "queryId is null");
+        requireNonNull(encoding, "encoding is null");
+        if (rows < 0) {
+            throw new IllegalArgumentException("rows is negative");
+        }
+        if (size < 0) {
+            throw new IllegalArgumentException("size is negative");
+        }
     }
 }
