+++ core/trino-spi/src/main/java/io/trino/spi/spool/SpoolingManager.java
@@ -11,10 +11,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.spi.protocol;
+package io.trino.spi.spool;
 
 import io.airlift.slice.Slice;
-import io.trino.spi.protocol.SpooledLocation.DirectLocation;
+import io.trino.spi.spool.SpooledLocation.DirectLocation;
 
 import java.io.IOException;
 import java.io.InputStream;
