+++ core/trino-spi/src/main/java/io/trino/spi/spool/SpoolingManagerFactory.java
@@ -11,21 +11,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.spi.protocol;
+package io.trino.spi.spool;
 
 import io.trino.spi.Experimental;
 
-import java.time.Instant;
+import java.util.Map;
 
-/**
- * SpooledSegmentHandle is used to uniquely identify a spooled segment and to manage its lifecycle.
- */
 @Experimental(eta = "2025-05-31")
-public interface SpooledSegmentHandle
+public interface SpoolingManagerFactory
 {
-    String identifier();
+    String getName();
 
-    String encoding();
-
-    Instant expirationTime();
+    SpoolingManager create(Map<String, String> config, SpoolingManagerContext context);
 }
