+++ plugin/trino-spooling-filesystem/src/main/java/io/trino/spooling/filesystem/FileSystemSpoolingManager.java
@@ -26,12 +26,12 @@ import io.trino.filesystem.TrinoFileSystemFactory;
 import io.trino.filesystem.TrinoInputFile;
 import io.trino.filesystem.TrinoOutputFile;
 import io.trino.filesystem.encryption.EncryptionKey;
-import io.trino.spi.protocol.SpooledLocation;
-import io.trino.spi.protocol.SpooledLocation.DirectLocation;
-import io.trino.spi.protocol.SpooledSegmentHandle;
-import io.trino.spi.protocol.SpoolingContext;
-import io.trino.spi.protocol.SpoolingManager;
 import io.trino.spi.security.ConnectorIdentity;
+import io.trino.spi.spool.SpooledLocation;
+import io.trino.spi.spool.SpooledLocation.DirectLocation;
+import io.trino.spi.spool.SpooledSegmentHandle;
+import io.trino.spi.spool.SpoolingContext;
+import io.trino.spi.spool.SpoolingManager;
 import io.trino.spooling.filesystem.encryption.EncryptionHeadersTranslator;
 import io.trino.spooling.filesystem.encryption.ExceptionMappingInputStream;
 
@@ -48,7 +48,7 @@ import java.util.concurrent.ThreadLocalRandom;
 import static io.airlift.slice.SizeOf.SIZE_OF_BYTE;
 import static io.airlift.slice.SizeOf.SIZE_OF_SHORT;
 import static io.trino.filesystem.encryption.EncryptionKey.randomAes256;
-import static io.trino.spi.protocol.SpooledLocation.coordinatorLocation;
+import static io.trino.spi.spool.SpooledLocation.coordinatorLocation;
 import static io.trino.spooling.filesystem.encryption.EncryptionHeadersTranslator.encryptionHeadersTranslator;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.time.Duration.between;
