+++ testing/trino-testing/src/main/java/io/trino/testing/LocalSpoolingManager.java
@@ -21,13 +21,13 @@ import com.google.errorprone.annotations.DoNotCall;
 import io.airlift.slice.Slice;
 import io.trino.client.JsonCodec;
 import io.trino.spi.Plugin;
-import io.trino.spi.protocol.SpooledLocation;
-import io.trino.spi.protocol.SpooledLocation.DirectLocation;
-import io.trino.spi.protocol.SpooledSegmentHandle;
-import io.trino.spi.protocol.SpoolingContext;
-import io.trino.spi.protocol.SpoolingManager;
-import io.trino.spi.protocol.SpoolingManagerContext;
-import io.trino.spi.protocol.SpoolingManagerFactory;
+import io.trino.spi.spool.SpooledLocation;
+import io.trino.spi.spool.SpooledLocation.DirectLocation;
+import io.trino.spi.spool.SpooledSegmentHandle;
+import io.trino.spi.spool.SpoolingContext;
+import io.trino.spi.spool.SpoolingManager;
+import io.trino.spi.spool.SpoolingManagerContext;
+import io.trino.spi.spool.SpoolingManagerFactory;
 import jakarta.annotation.PreDestroy;
 
 import java.io.IOException;
